ocr/requirements.txt
tensorflow==1.7.0
sanic==0.7.0
Pillow==5.1.0
